fix: split /game logic to their own files
